+++ routes/portal/instances.php
@@ -3,16 +3,14 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\Api\InstanceController;
 
 
-Route::middleware(['auth:api'])
+Route::middleware(['auth.api'])
     ->controller(InstanceController::class)
     ->prefix('catalog/instances')
     ->name('instances.')
     ->group(function () {
-        Route::get('/', 'index')->name('index');
-        Route::get('/{instanceId}', 'show')->name('show');
-        Route::post('/', 'store')->name('store');
-        Route::patch('/{instanceId}', 'update')->name('update');
-        Route::delete('/{instanceId}', 'destroy')->name('destroy');
-        // Instance routes
-        //Route::apiResource('', InstanceController::class);
+        Route::get('/', 'index')->name('index')->middleware('permission:instances.index');
+        Route::get('/{instanceId}', 'show')->name('show')->middleware('permission:instances.show');
+        Route::post('/', 'store')->name('store')->middleware('permission:instances.store');
+        Route::patch('/{instanceId}', 'update')->name('update')->middleware('permission:instances.update');
+        Route::delete('/{instanceId}', 'destroy')->name('destroy')->middleware('permission:instances.destroy');
     });
